[INTERNAL] warning removed
git-svn-id: http://svn.seqan.de/seqan/trunk@12747 e6417c60-b987-48fd-844e-b20f0fcc1017

+++ seqan/core/demos/tutorial/rnaseq/genequant_solution5.cpp
@@ -180,7 +180,7 @@ void outputGeneCoverage(String<unsigned> const & readBasesPerGene, TStore const 
 
             do
             {
-                mRNALength += abs(getAnnotation(exonIt).beginPos - getAnnotation(exonIt).endPos);
+                mRNALength += abs((int)getAnnotation(exonIt).beginPos - (int)getAnnotation(exonIt).endPos);
             }
             while (goRight(exonIt));
